+++ examples/github.rs
@@ -6,9 +6,7 @@ Example updating an executable to the latest version released via GitHub
 Include `signatures` in the features list to enable zipsign verification
 */
 
-// For the `cargo_crate_version!` macro
-#[macro_use]
-extern crate self_update;
+use self_update::cargo_crate_version;
 
 fn run() -> Result<(), Box<dyn ::std::error::Error>> {
     let mut rel_builder = self_update::backends::github::ReleaseList::configure();
@@ -44,6 +42,13 @@ fn run() -> Result<(), Box<dyn ::std::error::Error>> {
         // you obtain it via another mechanism, such as environment variables
         // or prompting the user for input
         //.auth_token(env!("DOWNLOAD_AUTH_TOKEN"))
+        //
+        // an optional `identifier` can be specified to narrow down
+        // an asset match for a target or OS-arch combination.
+        // additionally, if target/OS-arch search fails, then
+        // the identifier alone (if specified) will be used to
+        // find the asset
+        // .identifier("github-bin")
         .current_version(cargo_crate_version!())
         .build()?
         .update()?;