+++ src/lib.rs
@@ -1,8 +1,6 @@
-#![cfg_attr(feature = "cargo-clippy", deny(clippy::all))]
 /*!
 
 [![Build status](https://ci.appveyor.com/api/projects/status/xlkq8rd73cla4ixw/branch/master?svg=true)](https://ci.appveyor.com/project/jaemk/self-update/branch/master)
-[![Build Status](https://travis-ci.org/jaemk/self_update.svg?branch=master)](https://travis-ci.org/jaemk/self_update)
 [![crates.io:clin](https://img.shields.io/crates/v/self_update.svg?label=self_update)](https://crates.io/crates/self_update)
 [![docs](https://docs.rs/self_update/badge.svg)](https://docs.rs/self_update)
 
@@ -61,7 +59,8 @@ fn update() -> Result<(), Box<dyn std::error::Error>> {
 }
 ```
 
-Amazon S3, Google GCS, and DigitalOcean Spaces are also supported through the `S3` backend to check for new releases. Provided a `bucket_name`
+Amazon S3, Google GCS, and DigitalOcean Spaces, as well as any S3 compatible server are also supported
+through the `S3` backend to check for new releases.  Provided a `bucket_name`
 and `asset_prefix` string, `self_update` will look up all matching files using the following format
 as a convention for the filenames: `[directory/]<asset name>-<semver>-<platform/target>.<extension>`.
 Leading directories will be stripped from the file name allowing the use of subdirectories in the S3 bucket,
@@ -72,10 +71,13 @@ use self_update::cargo_crate_version;
 
 fn update() -> Result<(), Box<::std::error::Error>> {
     let status = self_update::backends::s3::Update::configure()
+        // .end_point(self_update::backends::s3::EndPoint::GCS)
+        // .end_point("https://s3.example.com")
         .bucket_name("self_update_releases")
         .asset_prefix("something/self_update")
         .region("eu-west-2")
         .bin_name("self_update_example")
+        // .access_key((env!("AWS_ACCESS_KEY_ID"), env!("AWS_SECRET_ACCESS_KEY")))
         .show_download_progress(true)
         .current_version(cargo_crate_version!())
         .build()?
@@ -136,15 +138,13 @@ self_update = { version = "0.27.0", features = ["rustls"], default-features = fa
 
 */
 
-#[macro_use]
-extern crate log;
 pub use self_replace;
 pub use tempfile::TempDir;
 
 #[cfg(feature = "compression-flate2")]
 use either::Either;
 use indicatif::{ProgressBar, ProgressStyle};
-use reqwest::header;
+use log::debug;
 use std::cmp::min;
 use std::fs;
 use std::io;
@@ -154,9 +154,12 @@ use std::path;
 mod macros;
 pub mod backends;
 pub mod errors;
+mod http_client;
 pub mod restart;
 pub mod update;
 pub mod version;
+
+use http_client::{header, HttpResponse};
 
 pub const DEFAULT_PROGRESS_TEMPLATE: &str =
     "[{elapsed_precise}] [{bar:40}] {bytes}/{total_bytes} ({eta}) {msg}";
@@ -375,6 +378,8 @@ impl<'a> Extract<'a> {
             None => detect_archive(self.source)?,
         };
 
+        // We cannot use a feature flag in a match arm. To bypass this the code block is
+        // isolated in a closure and called accordingly.
         let extract_into_plain_or_tar = |source: fs::File, compression: Option<Compression>| {
             let mut reader = Self::get_archive_reader(source, compression);
 
@@ -632,7 +637,7 @@ impl<'a> Move<'a> {
 pub struct Download {
     show_progress: bool,
     url: String,
-    headers: reqwest::header::HeaderMap,
+    headers: http_client::header::HeaderMap,
     progress_template: String,
     progress_chars: String,
 }
@@ -642,7 +647,7 @@ impl Download {
         Self {
             show_progress: false,
             url: url.to_owned(),
-            headers: reqwest::header::HeaderMap::new(),
+            headers: http_client::header::HeaderMap::new(),
             progress_template: DEFAULT_PROGRESS_TEMPLATE.to_string(),
             progress_chars: DEFAULT_PROGRESS_CHARS.to_string(),
         }
@@ -666,7 +671,7 @@ impl Download {
     }
 
     /// Set the download request headers, replaces the existing `HeaderMap`
-    pub fn set_headers(&mut self, headers: reqwest::header::HeaderMap) -> &mut Self {
+    pub fn set_headers(&mut self, headers: http_client::header::HeaderMap) -> &mut Self {
         self.headers = headers;
         self
     }
@@ -674,8 +679,8 @@ impl Download {
     /// Set a download request header, inserts into the existing `HeaderMap`
     pub fn set_header(
         &mut self,
-        name: reqwest::header::HeaderName,
-        value: reqwest::header::HeaderValue,
+        name: http_client::header::HeaderName,
+        value: http_client::header::HeaderValue,
     ) -> &mut Self {
         self.headers.insert(name, value);
         self
@@ -704,14 +709,11 @@ impl Download {
         }
 
         set_ssl_vars!();
-        let client = reqwest::blocking::ClientBuilder::new()
-            .use_rustls_tls()
-            .http2_adaptive_window(true)
-            .build()?;
-        let resp = client.get(&self.url).headers(headers).send()?;
+
+        let resp = http_client::get(&self.url, headers)?;
         let size = resp
             .headers()
-            .get(reqwest::header::CONTENT_LENGTH)
+            .get(http_client::header::CONTENT_LENGTH)
             .map(|val| {
                 val.to_str()
                     .map(|s| s.parse::<u64>().unwrap_or(0))
@@ -727,7 +729,7 @@ impl Download {
         }
         let show_progress = if size == 0 { false } else { self.show_progress };
 
-        let mut src = io::BufReader::new(resp);
+        let mut src = io::BufReader::new(resp.body());
         let mut downloaded = 0;
         let mut bar = if show_progress {
             let pb = ProgressBar::new(size);
@@ -767,6 +769,9 @@ impl Download {
 
 #[cfg(test)]
 mod tests {
+    #![allow(dead_code, unused_mut, unused_variables)]
+    // #![warn(unused_mut)]
+
     use super::*;
     #[cfg(feature = "compression-flate2")]
     use flate2::{self, write::GzEncoder};