+++ src/update.rs
@@ -1,11 +1,11 @@
 use regex::Regex;
-use reqwest::{self, header};
 use std::borrow::Cow;
 use std::env::consts::{ARCH, OS};
 use std::fs;
 use std::path::PathBuf;
 
-use crate::{confirm, errors::*, version, Download, Extract, Status};
+use crate::http_client::{self, header};
+use crate::{confirm, errors::*, version, Download, Extract, Move, Status};
 
 /// Release asset information
 #[derive(Clone, Debug, Default)]
@@ -63,14 +63,29 @@ impl Release {
     pub fn asset_for(&self, target: &str, identifier: Option<&str>) -> Option<ReleaseAsset> {
         self.assets
             .iter()
+            // first look specifically for a target with identifier
             .find(|asset| {
-                (asset.name.contains(target)
-                    || (asset.name.contains(OS) && asset.name.contains(ARCH)))
+                asset.name.contains(target)
                     && if let Some(i) = identifier {
                         asset.name.contains(i)
                     } else {
                         true
                     }
+            })
+            // otherwise look for a target for the current arch/os with identifier
+            .or_else(|| {
+                self.assets.iter().find(|asset| {
+                    (asset.name.contains(OS) && asset.name.contains(ARCH))
+                        && if let Some(i) = identifier {
+                            asset.name.contains(i)
+                        } else {
+                            true
+                        }
+                })
+            })
+            // otherwise just with the identifier if set
+            .or_else(|| {
+                identifier.and_then(|i| self.assets.iter().find(|asset| asset.name.contains(i)))
             })
             .cloned()
     }
@@ -128,14 +143,14 @@ pub trait ReleaseUpdate {
     /// Authorisation token for communicating with backend
     fn auth_token(&self) -> Option<String>;
 
-    /// ed25519ph verifying keys to validate a download's authenticy
+    /// ed25519ph verifying keys to validate a download's authenticity
     #[cfg(feature = "signatures")]
     fn verifying_keys(&self) -> &[[u8; zipsign_api::PUBLIC_KEY_LENGTH]] {
         &[]
     }
 
     /// Construct a header with an authorisation entry if an auth token is provided
-    fn api_headers(&self, auth_token: &Option<String>) -> Result<header::HeaderMap> {
+    fn api_headers(&self, auth_token: &Option<String>) -> Result<http_client::HeaderMap> {
         let mut headers = header::HeaderMap::new();
 
         if auth_token.is_some() {
@@ -211,6 +226,7 @@ pub trait ReleaseUpdate {
                             );
                             release.clone()
                         } else {
+                            println(show_output, "up-to-date.");
                             return Ok(UpdateStatus::UpToDate);
                         }
                     }
@@ -305,7 +321,12 @@ pub trait ReleaseUpdate {
         println(show_output, "Done");
 
         print_flush(show_output, "Replacing binary file... ")?;
-        self_replace::self_replace(new_exe)?;
+
+        if bin_install_path == std::env::current_exe()? {
+            self_replace::self_replace(new_exe)?;
+        } else {
+            Move::from_source(new_exe.as_ref()).to_dest(bin_install_path.as_ref())?;
+        }
         println(show_output, "Done");
 
         Ok(UpdateStatus::Updated(release))